fleet-manager/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
kubernetes
redis
httpx
psycopg2-binary
pydantic_settings